Resolvendo bug de quando clicava em voltar no carrinho
